+++ controller/konnect/watch_credentialacl.go
@@ -46,6 +46,9 @@ func kongCredentialACLReconciliationWatchOptions(
 				handler.EnqueueRequestsFromMapFunc(
 					kongCredentialACLForKongConsumer(cl),
 				),
+				builder.WithPredicates(
+					predicate.NewPredicateFuncs(objRefersToKonnectGatewayControlPlane[configurationv1.KongConsumer]),
+				),
 			)
 		},
 		func(b *ctrl.Builder) *ctrl.Builder {